Now the ConnectionAPI accepts a base_path again due to file paths problems, added some error handling in send and update file, now the API detects when it's sending a login command and save username and password for later usage, fixed tests in testMain -INCLUDING FILEWATCHER TEST- in order to add base_path when constructing a Connection object.
In particular:
The API now cleans the path that has passed to it using the base path OR complete the path in the case of the get_file method.

Now everytime we send a string the method uses a regex to check if the string (command) begins with "login ", in this case it tokenize the string and save username and password.

Implemented error handling due to errors during the sending of a file, when some chunks has already been sent. In this case the API closes the main socket, since the server cannot distinguish a message from the content of a file, and opens a new main socket. Once the socket has been connected again, the API automatically login again using the username and password saved in order to be ready to send new commands and files.